refactor(mode): move tmux slice under src/mode/tmux
git mv src/tmux{.rs,} -> src/mode/tmux{.rs,}; add src/mode.rs; repoint
crate::tmux::* -> crate::mode::tmux::*. No behavior change.
