+++ src/mode/tmux/locale.rs
