+++ src/mode/tmux/mouse/effect.rs
