+++ src/mode/tmux/paint_rescue.rs
