+++ src/mode/tmux/render.rs
@@ -2,8 +2,8 @@
 //! GPU render bundle to each projected `TmuxPane`, then routes tmux `%output`
 //! into the handle. Lives in the binary so `ozmux_tmux` stays renderer-free.
 
+use crate::mode::tmux::mode_ui::WorkspaceUiRoot;
 use crate::theme;
-use crate::tmux::mode_ui::WorkspaceUiRoot;
 use bevy::ecs::message::MessageReader;
 use bevy::math::Rect;
 use bevy::prelude::*;
